Back-end

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 import { MongooseError } from "mongoose";
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   // Video.find({}, (error, videos) => { //Callback Version
@@ -21,7 +22,8 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   // const id = req.params.id;
   const { id } = req.params; //ES6 Version
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  // const owner = await User.findById(video.owner); //populate를 사용하지 않으면 DB에 두번 접근
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
@@ -34,6 +36,9 @@ export const getEdit = async (req, res) => {
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
@@ -43,6 +48,9 @@ export const postEdit = async (req, res) => {
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -57,13 +65,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const file = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl: file.path,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -75,7 +92,13 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  console.log(id);
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(req.session.user._id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
